+++ virtual time capsule/src/main/java/com/java/web/virtual/time/capsule/repository/CapsuleRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CapsuleRepository extends JpaRepository<CapsuleEntity, Integer> {
+public interface CapsuleRepository extends JpaRepository<CapsuleEntity, Long> {
 }
